# go-web/C2-TT/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      emisor:
        type: string
      moneda:
        type: string
      monto:
        type: number
      receptor:
        type: string
      transaction_code:
        type: string
      transaction_date:
        type: string
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle MELI Transactions.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /transactions:
    get:
      description: get transactions
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List transactions
      tags:
      - Transactions
    patch:
      consumes:
      - application/json
      description: update transaction code and amount
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Transaction ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Code and amount to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: UpdateCodeAndAmount transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: store transactions
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Transaction to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store transactions
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: update transactions, providing all values
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: All transaction values to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update transactions
      tags:
      - Transactions
  /transactions{id}:
    delete:
      description: Delete transactions, providing id to delete
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete transaction
      tags:
      - Transactions
swagger: "2.0"